Modularize the search component further
