+++ src/components/Search/SearchInput.tsx
@@ -1,0 +1,18 @@
+import { ReactElement } from "react";
+import { Input, Stack } from "@chakra-ui/react";
+
+import { SearchInputProps } from "../../../typings/SearchInputProps";
+
+const SearchInput = ({ query, setQuery }: SearchInputProps): ReactElement => {
+    return (
+        <Stack direction={{ base: "column", sm: "row" }} gap={3}>
+            <Input
+                value={query}
+                onChange={(event) => setQuery(event.target.value)}
+                placeholder="Search movie title..."
+            />
+        </Stack>
+    );
+};
+
+export default SearchInput;
