+++ typings/SearchInputProps.ts
@@ -1,0 +1,4 @@
+export type SearchInputProps = {
+    query: string;
+    setQuery: (query: string) => void;
+};
